BUGFIX: prevent endless loop while deleting nodes
In our case (Neos 8.3), it happened that when a node was deleted,
the NodeReplicator activated, which then called $nodeVariant->remove(),
which again activated an nodeRemoved signal, which then activated the
NodeReplicator with the same node again (etc etc).

I am unsure why we did not catch this earlier (we are using NodeReplicator
since quite a long time already).